requirements.txt
asgiref==3.9.2
beautifulsoup4==4.14.2
crispy-bootstrap4==2025.6
Django==5.2.7
django-bootstrap4==25.2
django-crispy-forms==2.4
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
